support exclusive eni in shared eni

+++ init.sh
@@ -5,14 +5,18 @@ set -o nounset
 
 # install CNIs
 cp -f /usr/bin/terway /opt/cni/bin/
-cp -f /usr/bin/cilium-cni /opt/cni/bin/
 chmod +x /opt/cni/bin/terway
-chmod +x /opt/cni/bin/cilium-cni
+
+if [ "$TERWAY_DAEMON_MODE" != "VPC" ]; then
+  cp -f /usr/bin/cilium-cni /opt/cni/bin/
+  chmod +x /opt/cni/bin/cilium-cni
+fi
 
 # init cni config
 cp /tmp/eni/eni_conf /etc/eni/eni.json
 
 terway-cli cni /tmp/eni/10-terway.conflist /tmp/eni/10-terway.conf --output /etc/cni/net.d/10-terway.conflist
+terway-cli nodeconfig
 
 node_capabilities=/var/run/eni/node_capabilities
 if [ ! -f "$node_capabilities" ]; then
@@ -42,4 +46,7 @@ cp $node_capabilities /var-run-eni/node_capabilities
 
 sysctl -w net.ipv4.conf.eth0.rp_filter=0
 modprobe sch_htb || true
-chroot /host sh -c "systemctl disable eni.service; rm -f /etc/udev/rules.d/75-persistent-net-generator.rules /lib/udev/rules.d/60-net.rules /lib/udev/rules.d/61-eni.rules /lib/udev/write_net_rules"
+
+if [ "$TERWAY_DAEMON_MODE" != "VPC" ]; then
+  chroot /host sh -c "systemctl disable eni.service; rm -f /etc/udev/rules.d/75-persistent-net-generator.rules /lib/udev/rules.d/60-net.rules /lib/udev/rules.d/61-eni.rules /lib/udev/write_net_rules"
+fi
